Migrate Golangci-Lint configuration file
The steps suggested in the [migration guide][] cause too many
unecessary changes in the diff. The configuration file was therefor
migrated manually based on the automated migration in order to retain
any comments and the order of the settings.

`typecheck` is the only linter that has been removed, as it is no
longer considered a linter in version 2.

[migration guide]: https://golangci-lint.run/product/migration-guide/