move EnvPrefixGitLabRunner constant to gitlab namespace
both gitlab.EnvPrefixGitLabRunner and tart.EnvPrefixTartExecutor
are public now for consistency sake. initializing the tart
configuration requires passing a key prefix, which is
usually gitlab.EnvPrefixGitLabRunner. providing the prefix
ensures the gitlab and tart namespace remain independent.

additionally a proxy method for os.LookupEnv has been
implemented utilizing the newly added constant as prefix
for variable lookups.

this change is in preparation for #113